requirements.txt
streamlit>=1.30.0
google-generativeai>=0.5.0
firebase-admin>=6.4.0
Pillow>=10.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
twilio>=8.0.0
